+++ hospitapp/src/config/env.ts
@@ -10,6 +10,7 @@ interface EnvironmentVariables {
   NEXT_PUBLIC_APP_NAME: string;
   DATABASE_URL: string;
   DATABASE_NAME: string;
+  CACHE_TTL: number;
 }
 
 /**
@@ -25,6 +26,7 @@ const validate_env = (): EnvironmentVariables => {
     NEXT_PUBLIC_APP_NAME: process.env.NEXT_PUBLIC_APP_NAME,
     DATABASE_URL: process.env.DATABASE_URL,
     DATABASE_NAME: process.env.DATABASE_NAME,
+    CACHE_TTL: parseInt(process.env.CACHE_TTL || '86400', 10)
   };
 
   // Check for missing required variables
