+++ hospitapp/src/services/search_ips/search_ips_mongo.service.ts
